requirements.txt
Django>=4.2
djangorestframework>=3.14
gunicorn>=21.2
dj-database-url>=1.0
psycopg2-binary>=2.9
python-decouple>=3.8
whitenoise>=6.5
django-cors-headers>=4.0
Pillow>=10.0        # if using ImageField
django-filter>=23.1 # if using DRF filters